+++ CHANGELOG.md
@@ -79,6 +79,7 @@ Putting this stuff down here so things don't conflict as often.
 * Added `TSPlayer` to `GetDataHandlers.PlayerAnimation`. (@hakusaro)
 * Added `GetDataHandlers.MassWireOperation` hook and related arguments. (@hakusaro)
 * Added `GetDataHandlers.PlaceTileEntity` hook and related arguments. (@hakusaro)
+* Added `TSPlayer` to `GetDataHandlers.GemLockToggle`. (@hakusaro)
 
 ## TShock 4.3.25
 * Fixed a critical exploit in the Terraria protocol that could cause massive unpreventable world corruption as well as a number of other problems. Thanks to @bartico6 for reporting. Fixed by the efforts of @QuiCM, @hakusaro, and tips in the right directioon from @bartico6.